cirrus-apps/requirements.txt
flask
flask_session
pyyaml